Code stability improvements and new features implemented

+++ src/modules/pages/api/apiPageDelete.js
@@ -2,19 +2,21 @@ import {
     ObjectId
 } from "mongodb";
 import pageDelete from "./data/pageDelete.json";
-import Auth from "../../../shared/lib/auth";
-import C from "../../../shared/lib/constants";
 
 export default () => ({
     schema: {
         body: pageDelete.root
     },
     attachValidation: true,
-    async handler(req, rep) {
-        const response = new this.Response(req, rep); const log = new this.LoggerHelpers(req, this);
+    async handler(req) {
+        const {
+            log,
+            response,
+            auth,
+            acl
+        } = req.zoia;
         // Check permissions
-        const auth = new Auth(this.mongo.db, this, req, rep, C.USE_BEARER_FOR_TOKEN);
-        if (!(await auth.getUserData()) || !auth.checkStatus("admin")) {
+        if (!auth.checkStatus("admin")) {
             response.unauthorizedError();
             return;
         }
@@ -25,6 +27,34 @@ export default () => ({
             return;
         }
         try {
+            // Build query
+            const queryDb = {
+                $or: req.body.ids.map(id => ({
+                    _id: new ObjectId(id)
+                }))
+            };
+            // Get requested data
+            const dataDb = await this.mongo.db.collection(req.zoiaModulesConfig["pages"].collectionPages).find(queryDb, {
+                projection: {
+                    filename: 1
+                }
+            }).toArray();
+            // Check permission
+            let allowed = true;
+            (dataDb || []).map(i => {
+                if (allowed && !acl.checkPermission("pages", "delete", i.filename)) {
+                    allowed = false;
+                }
+            });
+            if (!allowed) {
+                response.requestError({
+                    failed: true,
+                    error: "Access Denied",
+                    errorKeyword: "accessDenied",
+                    errorData: []
+                });
+                return;
+            }
             // Delete requested IDs
             const result = await this.mongo.db.collection(req.zoiaModulesConfig["pages"].collectionPages).deleteMany({
                 $or: req.body.ids.map(id => ({
